airplay-multiroom-server/requirements.txt
flask>=2.3.0
python-zeroconf>=0.39.0
pygobject>=3.42.0
pycairo>=1.20.0
pyyaml>=6.0
asyncio>=3.4.3
aiohttp>=3.8.0
websockets>=11.0
cryptography>=3.4.8
netifaces>=0.11.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=4.0.0